refactor: move mapgroup into util
