+++ core/src/main/java/dk/gtz/graphedit/view/ModelEditorController.java
@@ -13,6 +13,7 @@ import dk.gtz.graphedit.tool.IToolbox;
 import dk.gtz.graphedit.events.ViewportKeyEvent;
 import dk.gtz.graphedit.events.ViewportMouseEvent;
 import dk.gtz.graphedit.spi.ISyntaxFactory;
+import dk.gtz.graphedit.util.MapGroup;
 import dk.gtz.graphedit.util.MetadataUtils;
 import dk.gtz.graphedit.viewmodel.IFocusable;
 import dk.gtz.graphedit.viewmodel.ViewModelEdge;
